# internal/backend/openapi.gen.go
// Package backend provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package backend

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for WithdrawFlowItemEventType.
const (
	WithdrawFlowItemEventTypeRequest  WithdrawFlowItemEventType = "request"
	WithdrawFlowItemEventTypeWithdraw WithdrawFlowItemEventType = "withdraw"
)

// Defines values for ListWithdrawFlowsParamsEventType.
const (
	ListWithdrawFlowsParamsEventTypeRequest  ListWithdrawFlowsParamsEventType = "request"
	ListWithdrawFlowsParamsEventTypeWithdraw ListWithdrawFlowsParamsEventType = "withdraw"
)

// ApiError defines model for ApiError.
type ApiError struct {
	Error string `json:"error"`
}

// DailySeriesDataPoint defines model for DailySeriesDataPoint.
type DailySeriesDataPoint struct {
	Day          openapi_types.Date `json:"day"`
	RequestCount float64            `json:"request_count"`
	Requested    float64            `json:"requested"`
	Withdrawn    float64            `json:"withdrawn"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// ListWithdrawFlowsMeta defines model for ListWithdrawFlowsMeta.
type ListWithdrawFlowsMeta struct {
	Count  int `json:"count"`
	Limit  int `json:"limit"`
	Offset int `json:"offset"`
}

// ListWithdrawFlowsResponse defines model for ListWithdrawFlowsResponse.
type ListWithdrawFlowsResponse struct {
	Items []WithdrawFlowItem    `json:"items"`
	Meta  ListWithdrawFlowsMeta `json:"meta"`
}

// WithdrawFlowItem defines model for WithdrawFlowItem.
type WithdrawFlowItem struct {
	AmountNormalized string                    `json:"amount_normalized"`
	AmountRaw        string                    `json:"amount_raw"`
	BlockNumber      int64                     `json:"block_number"`
	BlockTime        time.Time                 `json:"block_time"`
	ChainId          int64                     `json:"chain_id"`
	EventType        WithdrawFlowItemEventType `json:"event_type"`
	LogIndex         int32                     `json:"log_index"`
	SenderAddress    string                    `json:"sender_address"`
	TxHash           string                    `json:"tx_hash"`
	UpdatedAt        time.Time                 `json:"updated_at"`
}

// WithdrawFlowItemEventType defines model for WithdrawFlowItem.EventType.
type WithdrawFlowItemEventType string

// ListWithdrawFlowsParams defines parameters for ListWithdrawFlows.
type ListWithdrawFlowsParams struct {
	// SenderAddress Filter by sender address.
	SenderAddress *string `form:"sender_address,omitempty" json:"sender_address,omitempty"`

	// EventType Filter by cooldown event type.
	EventType *ListWithdrawFlowsParamsEventType `form:"event_type,omitempty" json:"event_type,omitempty"`

	// FromTime Include rows with block time >= this timestamp (RFC3339).
	FromTime *time.Time `form:"from_time,omitempty" json:"from_time,omitempty"`

	// ToTime Include rows with block time <= this timestamp (RFC3339).
	ToTime *time.Time `form:"to_time,omitempty" json:"to_time,omitempty"`

	// Limit Maximum number of items. Default 100, capped at 500.
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Number of rows to skip.
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// ListWithdrawFlowsParamsEventType defines parameters for ListWithdrawFlows.
type ListWithdrawFlowsParamsEventType string

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List daily chart series points
	// (GET /daily-series-data)
	ListDailySeriesData(w http.ResponseWriter, r *http.Request)
	// Health check
	// (GET /healthz)
	GetHealth(w http.ResponseWriter, r *http.Request)
	// List raw withdrawal flow events
	// (GET /withdraw-flows)
	ListWithdrawFlows(w http.ResponseWriter, r *http.Request, params ListWithdrawFlowsParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// List daily chart series points
// (GET /daily-series-data)
func (_ Unimplemented) ListDailySeriesData(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Health check
// (GET /healthz)
func (_ Unimplemented) GetHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List raw withdrawal flow events
// (GET /withdraw-flows)
func (_ Unimplemented) ListWithdrawFlows(w http.ResponseWriter, r *http.Request, params ListWithdrawFlowsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListDailySeriesData operation middleware
func (siw *ServerInterfaceWrapper) ListDailySeriesData(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListDailySeriesData(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListWithdrawFlows operation middleware
func (siw *ServerInterfaceWrapper) ListWithdrawFlows(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListWithdrawFlowsParams

	// ------------- Optional query parameter "sender_address" -------------

	err = runtime.BindQueryParameter("form", true, false, "sender_address", r.URL.Query(), &params.SenderAddress)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "sender_address", Err: err})
		return
	}

	// ------------- Optional query parameter "event_type" -------------

	err = runtime.BindQueryParameter("form", true, false, "event_type", r.URL.Query(), &params.EventType)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "event_type", Err: err})
		return
	}

	// ------------- Optional query parameter "from_time" -------------

	err = runtime.BindQueryParameter("form", true, false, "from_time", r.URL.Query(), &params.FromTime)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "from_time", Err: err})
		return
	}

	// ------------- Optional query parameter "to_time" -------------

	err = runtime.BindQueryParameter("form", true, false, "to_time", r.URL.Query(), &params.ToTime)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "to_time", Err: err})
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", r.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", r.URL.Query(), &params.Offset)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "offset", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListWithdrawFlows(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/daily-series-data", wrapper.ListDailySeriesData)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/healthz", wrapper.GetHealth)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/withdraw-flows", wrapper.ListWithdrawFlows)
	})

	return r
}